create models pedidos, produtos, usuario and respectives migrations and resources

+++ src/database/migrations/20230513202249-create-produtos.js
@@ -15,7 +15,7 @@ module.exports = {
       },
       valor: {
         allowNull: false,
-        type: Sequelize.DECIMAL(10,2)
+        type: Sequelize.FLOAT
       },
       categoria: {
         allowNull: false,
